ballbotgym/pyproject.toml
# This is the pyproject.toml for the ballbotgym package.
# 
# Documentation notes:
# - This package provides a MuJoCo-based reinforcement learning environment for the Ballbot robot.
# - The environment is compatible with Gymnasium (gymnasium.make("ballbot-v0.1")).
# - Dependency on MuJoCo is *not* included because a patched version is required (see README for install instructions).
# - See ballbotgym/__init__.py and ballbotgym/bbot_env.py for main env registration/implementation.
# - Maintainer: [Nebiyu Tadesse]
#
# For more on the pyproject.toml format see: https://packaging.python.org/en/latest/specifications/declaring-project-metadata/

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ballbotgym"
version = "0.1.0"
# Required dependencies for core functionality:
dependencies = [
  # mujoco is intentionally omitted: requires a patched version, see README
  "gymnasium",
  "numpy-quaternion",
  "matplotlib",
  "scipy",
  "opencv-python",
  "termcolor"
]
# (you can add [project.optional-dependencies], [tool.*], etc. blocks here as needed)
